requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-crispy-forms==1.13.0
django-filter==21.1
django-nested-admin==3.4.0
django-nested-inline==0.4.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
mysqlclient==2.1.0
Pillow==9.0.0
PyJWT==1.7.1
python-memcached==1.59
python-monkey-business==1.0.0
python3-memcached==1.51
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
